handlers/cmcard/requirements.txt
python-telegram-bot[job-queue]==21.6
python-dotenv
razorpay
selenium
requests
beautifulsoup4
reportlab
pillow
PyPDF2
